refactor: Breakup calculator module. Extract ForceLevel, extend options and integrate directly with CLI. Implement tests.
